requirements.txt
cryptography>=43.0.0
pyyaml>=6.0.0
